로그인 채로 사이트 기능 용시, 이용시간 다시 30분으로 되도록 코드 수정 (테스트를위해 토큰시간 20초로 조정)

+++ app/src/main/java/bitcamp/carrot_thunder/chatting/model/dao/ChattingDAO.java
@@ -16,9 +16,9 @@ public interface ChattingDAO {
 
   void saveMessage(ChatMessageVO message, String sellerRoomId, String buyerRoomId);
 
-  List<ChatRoomVO> getChatRoomsForSeller(int sellerId);
+  List<ChatRoomVO> getChatRoomsForSeller(long sellerId);
 
-  List<ChatRoomVO> getChatRoomsForMember(int memberId);
+  List<ChatRoomVO> getChatRoomsForMember(long memberId);
 
   String checkChatRoomExists(@Param("sellerId") int sellerId,
       @Param("currentUserId") int currentUserId, @Param("postId") int postId, @Param("userId") int userId);
